+++ engine/crates/composition/CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## Unreleased
 
+- Added composition for default values of output field arguments and input fields. They are now reflected in the federated graph.
 - Support the experimental @authorized directive
+- Fixed the ingestion of numeric literals when creating a federated graph from a string.
 
 ## 0.4.0 - 2024-06-11